+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added restore confirmation dialogs ([#447])
+
 ## [1.2.1] - 2024-09-28
 
 ### Added